Challenge 2 completed

+++ src/estimator.py
@@ -27,15 +27,15 @@ def impact_cal(data={}):
         'currentlyInfected': data['reportedCases'] * 10,
         'infectionsByRequestedTime': impact['currentlyInfected'] * infectionsCalcution(data['periodType'],
                                                                                        data['timeToElapse']),
-        'severeCasesByRequestedTime': float(0.15 * impact['infectionsByRequestedTime'])
+        'severeCasesByRequestedTime': 0.15 * impact['infectionsByRequestedTime']
     }
 
     impact = {
         'currentlyInfected': data['reportedCases'] * 10,
         'infectionsByRequestedTime': impact['currentlyInfected'] * infectionsCalcution(data['periodType'],
                                                                                        data['timeToElapse']),
-        'severeCasesByRequestedTime': float(0.15 * impact['infectionsByRequestedTime']),
-        'hospitalBedsByRequestedTime': float(0.35 * data['totalHospitalBeds']) - impact['severeCasesByRequestedTime']
+        'severeCasesByRequestedTime': 0.15 * impact['infectionsByRequestedTime'],
+        'hospitalBedsByRequestedTime': 0.35 * data['totalHospitalBeds'] - impact['severeCasesByRequestedTime']
     }
 
     return impact
@@ -55,15 +55,15 @@ def severeImpact_cal(data={}):
         'currentlyInfected': data['reportedCases'] * 50,
         'infectionsByRequestedTime': severeImpact['currentlyInfected'] * infectionsCalcution(data['periodType'],
                                                                                              data['timeToElapse']),
-        'severeCasesByRequestedTime': float(0.15 * severeImpact['infectionsByRequestedTime'])
+        'severeCasesByRequestedTime': 0.15 * severeImpact['infectionsByRequestedTime']
     }
 
     severeImpact = {
         'currentlyInfected': data['reportedCases'] * 50,
         'infectionsByRequestedTime': severeImpact['currentlyInfected'] * infectionsCalcution(data['periodType'],
                                                                                              data['timeToElapse']),
-        'severeCasesByRequestedTime': float(0.15 * severeImpact['infectionsByRequestedTime']),
-        'hospitalBedsByRequestedTime': float(0.35 * data['totalHospitalBeds']) - severeImpact['severeCasesByRequestedTime']
+        'severeCasesByRequestedTime': 0.15 * severeImpact['infectionsByRequestedTime'],
+        'hospitalBedsByRequestedTime': 0.35 * data['totalHospitalBeds'] - severeImpact['severeCasesByRequestedTime']
     }
 
     return severeImpact
